gemm_optimized mixed dmr

+++ src/cuda/gemm_tensorcores/gemm_optimized/main.cpp
@@ -14,8 +14,8 @@
 #include <vector>
 #include <iostream>
 
-typedef float real_t;
-typedef float half_real_t;
+typedef double real_t;
+typedef double half_real_t;
 
 void gemm_host(std::vector<real_t>& a, std::vector<real_t>& b,
 		std::vector<real_t>& c, real_t alpha, real_t beta, int m, int n,
